Allow the router resolver to work with export default
If you're on a newer Ember version, then you'll be doing
`import EmberRouter from `@ember/routing/router'` rather
than calling the `Ember.Router` global directly. This means
you have to do an `export default` instead of `module.exports =`.

In that situation, `options.context(name)` returns the module,
and we need to call the default export on it.